+++ Assets/Scripts/Vehicle.cs
@@ -7,9 +7,9 @@ public class Vehicle : MonoBehaviour
     public string VehicleName;
     public string VehicleType;
     public int IsBought = 0;
-    public double VehiclePrice;
-    public double VehicleProduceSpeed;
-    public double VehicleProduceAmount;
-    public double VehicleUpgradeCost;
+    public float VehiclePrice;
+    public float VehicleProduceSpeed;
+    public float VehicleProduceAmount;
+    public float VehicleUpgradeCost;
 
 }
